tests/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
prometheus-client>=0.16.0
pyyaml>=6.0.0
requests>=2.28.0
redis>=5.0.1
